fix SSR test widget setup and harness stylesheet
