+++ packages/fast-test-harness/test/index.html
@@ -3,6 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <title>Fast Test Harness — Self Test</title>
+        <link rel="stylesheet" href="/test-theme.css" />
     </head>
     <body>
         <script type="module" src="/src/main.ts"></script>
