+++ packages/fast-test-harness/test/public/test-theme.css
@@ -1,5 +1,5 @@
 :root {
-    --test-theme-color: red;
+    --test-theme-color: orange;
 }
 
 * {
