+++ packages/fast-test-harness/test/src/entry-client.ts
@@ -1,6 +1,7 @@
-import { RenderableFASTElement } from "@microsoft/fast-html";
 import "@microsoft/fast-test-harness/ssr/entry-client.js";
-import { definition, TestWidget } from "./main.js";
+
+import { RenderableFASTElement } from "@microsoft/fast-html";
+import { definition, TestWidget } from "./test-widget/test-widget.js";
 
 RenderableFASTElement(TestWidget).defineAsync({
     name: definition.name,
