+++ packages/fast-test-harness/test/src/test-widget/test-widget.styles.css
@@ -1,0 +1,7 @@
+:host {
+    display: inline-block;
+
+    .label {
+        font-weight: bold;
+    }
+}
