+++ packages/fast-test-harness/test/src/test-widget/test-widget.template.html
@@ -1,0 +1,6 @@
+<f-template name="test-widget" shadowrootmode="open">
+    <template>
+        <span class="label">{{label}}</span>
+        <slot></slot>
+    </template>
+</f-template>
